Create Details page for Articles and Projects

+++ DevCard_MVC/Models/Project.cs
@@ -7,14 +7,14 @@ namespace DevCard_MVC.Models
 {
     public class Project
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public  string Description { get; set; }
         public  string Client { get; set;}
         public string ProjectImage { get; set; }
 
 
-        public Project(int id, string name, string description, string client, string projectImage)
+        public Project(long id, string name, string description, string client, string projectImage)
         {
             Id = id;
             Name = name;
